// src/server/rpc/generated/schemaforge/v1/SchemaSummary.ts
// Original file: schema.proto


export interface SchemaSummary {
  'id'?: (string);
  'name'?: (string);
  'description'?: (string);
  'entityCount'?: (number);
  'relationshipCount'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
}

export interface SchemaSummary__Output {
  'id': (string);
  'name': (string);
  'description': (string);
  'entityCount': (number);
  'relationshipCount': (number);
  'createdAt': (string);
  'updatedAt': (string);
}